+++ Clusterlib/Vector.cs
@@ -9,16 +9,16 @@ namespace ClusterLib
     [Serializable()]
     public class Vector
     {
-        public decimal[] vec = new decimal[3]; //Field
+        public double[] vec = new double[3]; //Field
 
         public Vector() { }         //blank constructor
 
-        public Vector(decimal[] vec) //decimal[3] overload
+        public Vector(double[] vec) //double[3] overload
         {
             this.vec = vec;
         }
 
-        public Vector(decimal a,decimal b,decimal c) //decimal[3] overload
+        public Vector(double a,double b,double c) //double[3] overload
         {
             this.vec[0] = a;
             this.vec[1] = b;
@@ -38,27 +38,27 @@ namespace ClusterLib
 
         public static Vector operator+ (Vector a, Vector b)
         {
-            return new Vector(new decimal[] { a.vec[0] + b.vec[0], a.vec[1] + b.vec[1], a.vec[2] + b.vec[2] });
+            return new Vector(new double[] { a.vec[0] + b.vec[0], a.vec[1] + b.vec[1], a.vec[2] + b.vec[2] });
         }
 
         public static Vector operator -(Vector a, Vector b)
         {
-            return new Vector(new decimal[] { a.vec[0] - b.vec[0], a.vec[1] - b.vec[1], a.vec[2] - b.vec[2] });
+            return new Vector(new double[] { a.vec[0] - b.vec[0], a.vec[1] - b.vec[1], a.vec[2] - b.vec[2] });
         }
 
         public static Vector operator *(Vector a, Vector b)
         {
-            return new Vector(new decimal[] { a.vec[0] * b.vec[0], a.vec[1] * b.vec[1], a.vec[2] * b.vec[2] });
+            return new Vector(new double[] { a.vec[0] * b.vec[0], a.vec[1] * b.vec[1], a.vec[2] * b.vec[2] });
         }
 
         public static Vector operator /(Vector a, Vector b)
         {
-            return new Vector(new decimal[] { a.vec[0] / b.vec[0], a.vec[1] / b.vec[1], a.vec[2] / b.vec[2] });
+            return new Vector(new double[] { a.vec[0] / b.vec[0], a.vec[1] / b.vec[1], a.vec[2] / b.vec[2] });
         }
 
-        public static Vector operator *(decimal a, Vector b)
+        public static Vector operator *(double a, Vector b)
         {
-            return new Vector(new decimal[] { a * b.vec[0], a* b.vec[1], a* b.vec[2] });
+            return new Vector(new double[] { a * b.vec[0], a* b.vec[1], a* b.vec[2] });
         }
 
 
@@ -78,40 +78,40 @@ namespace ClusterLib
             for (int i = 0; i < 3; i++)
                 this.vec[i] -= vec.vec[i];
         }
-        public void mult(decimal value) //mult this with
+        public void mult(double value) //mult this with
         {
             for (int i = 0; i < 3; i++)
                 this.vec[i] *= value;
         }
-        public void div(decimal value)   //div this by
+        public void div(double value)   //div this by
         {
             for (int i = 0; i < 3; i++)
                 this.vec[i] /= value;
         }
-        public decimal skalar(Vector vec)    // skalarprodukt
+        public double skalar(Vector vec)    // skalarprodukt
         {
-            decimal output = 0;
+            double output = 0;
             for (int i = 0; i < 3; i++)
                 output += this.vec[i] * vec.vec[i];
             return output;
         }
-        public decimal distance()            //magnitude of the vector
+        public double distance()            //magnitude of the vector
         {
-            decimal hypo = 0;
+            double hypo = 0;
             for (int i = 0; i <= 2; i++)
                 hypo += this.vec[i]* this.vec[i];
-            return (decimal)Math.Sqrt((Double)hypo);
+            return (double)Math.Sqrt((Double)hypo);
         }
         public Vector direction(Vector vec2) //calc directionvector to other vector
         {
             
             return vec2 - this;
         }
-        public void scale(decimal distance)   //scale magnitude to value
+        public void scale(double distance)   //scale magnitude to value
         {
-            decimal d = distance / this.distance();
+            double d = distance / this.distance();
             for (int i = 0; i <= 2; i++)
-                this.vec[i] = (decimal)Math.Sqrt(Math.Pow((double)this.vec[i],2)*(double)d);
+                this.vec[i] = (double)Math.Sqrt(Math.Pow((double)this.vec[i],2)*(double)d);
         }
 
         public Vector random(double range=1)   //scale magnitude to value
@@ -123,9 +123,9 @@ namespace ClusterLib
 
         /*public Vector polar()
         {
-            decimal r = this.distance();
-            decimal g = 0;
-            decimal b = 0;
+            double r = this.distance();
+            double g = 0;
+            double b = 0;
             if (this.vec[0] > 0)
                 if (this.vec[1] > 0)
                     g = this.vec[1] / (Math.Abs(this.vec[0]) + Math.Sqrt(Math.Pow(this.vec[0], 2) + Math.Pow(this.vec[1], 2)));
@@ -135,8 +135,8 @@ namespace ClusterLib
                 g = 180 - this.vec[1] / (Math.Abs(this.vec[0]) + Math.Sqrt(Math.Pow(this.vec[0], 2) + Math.Pow(this.vec[1], 2)));
 
             if (r != 0)
-                b = (decimal)Math.Asin((double)this.vec[2] / r);
-            return new ClusterLib.Vector(new decimal[3] { r, Math.Asin(this.vec[2] / r), g });
+                b = (double)Math.Asin((double)this.vec[2] / r);
+            return new ClusterLib.Vector(new double[3] { r, Math.Asin(this.vec[2] / r), g });
 
         }*/
 
@@ -150,11 +150,11 @@ namespace ClusterLib
     }
     public class Vec6
     {
-        public decimal[] vec = new decimal[6]; //Field
+        public double[] vec = new double[6]; //Field
 
         public Vec6() { }         //blank constructor
 
-        public Vec6(decimal[] vec) //decimal[6] overload
+        public Vec6(double[] vec) //double[6] overload
         {
             this.vec = vec;
         }
@@ -201,7 +201,7 @@ namespace ClusterLib
             return temp;
         }
 
-        public static Vec6 operator *(decimal a, Vec6 b)
+        public static Vec6 operator *(double a, Vec6 b)
         {
             Vec6 temp = new Vec6();
             for (int i = 0; i < 6; i++)
@@ -226,29 +226,29 @@ namespace ClusterLib
             for (int i = 0; i < 6; i++)
                 this.vec[i] -= vec.vec[i];
         }
-        public void mult(decimal value) //mult this with
+        public void mult(double value) //mult this with
         {
             for (int i = 0; i < 6; i++)
                 this.vec[i] *= value;
         }
-        public void div(decimal value)   //div this by
+        public void div(double value)   //div this by
         {
             for (int i = 0; i < 6; i++)
                 this.vec[i] /= value;
         }
-        public decimal skalar(Vec6 vec)    // skalarprodukt
+        public double skalar(Vec6 vec)    // skalarprodukt
         {
-            decimal output = 0;
+            double output = 0;
             for (int i = 0; i < 6; i++)
                 output += this.vec[i] * vec.vec[i];
             return output;
         }
-        public decimal distance()            //magnitude of the Vec6
+        public double distance()            //magnitude of the Vec6
         {
-            decimal hypo = 0;
+            double hypo = 0;
             for (int i = 0; i <= 5; i++)
-                hypo += (decimal)Math.Pow((double)this.vec[i], 5);
-            return (decimal)Math.Sqrt((double)hypo);
+                hypo += (double)Math.Pow((double)this.vec[i], 5);
+            return (double)Math.Sqrt((double)hypo);
         }
         public Vec6 direction(Vec6 vec5) //calc directionVec6 to other Vec6
         {
@@ -260,9 +260,9 @@ namespace ClusterLib
         public Vector ToVector(int i)
         {
             if (i == 0)
-                return new Vector(new decimal[] { this.vec[0], this.vec[1], this.vec[2] });
+                return new Vector(new double[] { this.vec[0], this.vec[1], this.vec[2] });
             else
-                return new Vector(new decimal[] { this.vec[3], this.vec[4], this.vec[5] });
+                return new Vector(new double[] { this.vec[3], this.vec[4], this.vec[5] });
         }
 
         public string toString()            //Translate to String
